+++ student/java/tenmo-server/src/main/java/com/techelevator/tenmo/model/User.java
@@ -6,7 +6,7 @@ import java.util.Set;
 
 public class User {
 
-   private Long id;
+   private Integer id;
    private String username;
    private String password;
    private boolean activated;
@@ -14,18 +14,18 @@ public class User {
 
    public User() { }
 
-   public User(Long id, String username, String password, String authorities) {
+   public User(Integer id, String username, String password, String authorities) {
       this.id = id;
       this.username = username;
       this.password = password;
       this.activated = true;
    }
 
-   public Long getId() {
+   public Integer getId() {
       return id;
    }
 
-   public void setId(Long id) {
+   public void setId(Integer id) {
       this.id = id;
    }
